+++ src/App.js
@@ -1,18 +1,13 @@
 import './styles/App.css';
-import MyNavbar from "./components/head/MyNavbar";
+import MyNavbar from "./components/navbar/MyNavbar";
 import MyMain from "./components/main/MyMain";
 
 function App() {
 
   return (
     <div className="App">
-        <MyNavbar>
-
-        </MyNavbar>
-
-        <MyMain>
-
-        </MyMain>
+        <MyNavbar />
+        <MyMain />
     </div>
   );
 }
