+++ src/components/UI/button/MyButton.module.css
@@ -12,3 +12,4 @@
 .myBtn:hover {
     color: #13bc3e;
 }
+
