+++ src/components/footer/MyFooter.jsx
@@ -1,6 +1,7 @@
 import React from 'react';
 
-const Footer = () => {
+
+const MyFooter = () => {
     return (
         <div>
 
@@ -8,4 +9,4 @@ const Footer = () => {
     );
 };
 
-export default Footer;
+export default MyFooter;
