+++ src/components/menu/MyMenu.module.css
@@ -7,5 +7,7 @@
 }
 
 
-
+.iconButton {
+    margin-right: 10px;
+}
